% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_x11.R
\name{henderson}
\alias{henderson}
\title{Apply Henderson linear filter}
\usage{
henderson(y, length, musgrave = TRUE, ic = 4.5)
}
\arguments{
\item{y}{input time-series.}

\item{length}{length of the Henderson filter.}

\item{musgrave}{Boolean indicating if Musgrave asymmetric filters should be used.}

\item{ic}{ic ratio: irregular/trend-cycle.}
}
\description{
Apply Henderson linear filter
}
