% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_x11.R
\name{x11plus}
\alias{x11plus}
\title{Perform an X-11 like decomposition with any (non integer) periodicity.}
\usage{
x11plus(
  y,
  period,
  mul = TRUE,
  trend.horizon = 6,
  trend.degree = 2,
  trend.kernel = c("Henderson", "BiWeight", "TriWeight", "TriCube", "Uniform",
    "Triangular", "Epanechnikov", "Trapezoidal"),
  trend.asymmetric = c("CutAndNormalize", "Direct", "MMSRE"),
  seas.s0 = c("S3X3", "S3X1", "S3X5", "S3X9", "S3X15"),
  seas.s1 = c("S3X5", "S3X3", "S3X1", "S3X9", "S3X15"),
  extreme.lsig = 1.5,
  extreme.usig = 2.5
)
}
\arguments{
\item{y}{input time-series.}

\item{period}{Period of the seasonal component, any positive real number.}

\item{mul}{Boolean indicating if the decomposition mode is multiplicative (TRUE).}

\item{trend.horizon}{bandwidth of trend filters.}

\item{trend.degree}{polynomial order in local trend model.}

\item{trend.kernel}{kernel weights in objective function.}

\item{trend.asymmetric}{truncation type for symmetric filter.}

\item{seas.s0}{Seasonal filter for B5, C5, D5.}

\item{seas.s1}{seasonal filter for B10, C10, D10.}

\item{extreme.lsig}{lower boundary used for outlier correction in irregular.}

\item{extreme.usig}{upper boundary used for outlier correction in irregular.}
}
\description{
Perform an X-11 like decomposition with any (non integer) periodicity.
}
