% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_x11.R
\name{henderson}
\alias{henderson}
\title{Apply Henderson linear filter}
\usage{
henderson(x, length, musgrave = TRUE, ic = 4.5)
}
\arguments{
\item{x}{input time-series.}

\item{length}{length of the Henderson filter.}

\item{musgrave}{Boolean indicating if Musgrave asymmetric filters should be used.}

\item{ic}{ic ratio: irregular/trend-cycle.}
}
\value{
A numeric array corresponding to the the trend
}
\description{
Apply Henderson linear filter
}
\examples{
q<-x11plus(rjd3toolkit::ABS$X0.2.09.10.M, 12)

henderson(q$decomposition$sa, 13)
}
